security tests are a suite, not a helper
