+++ packages/api-rest-handlers/__tests__/absorb.spec.js
@@ -1,7 +1,6 @@
 const { absorbHandler } = require('../absorb');
 
 const { setupMocks, neo4jTest } = require('../../../test-helpers');
-const { securityTests } = require('../../../test-helpers/security');
 const { dbUnavailable } = require('../../../test-helpers/error-stubs');
 
 describe('rest DELETE', () => {
@@ -42,8 +41,6 @@ describe('rest DELETE', () => {
 				),
 			],
 		);
-
-	securityTests(absorbHandler(), mainCode);
 
 	describe('error handling', () => {
 		it('responds with 500 if neo4j query fails', async () => {
